+++ src/net-util/export/IntfRequestHandler.js
@@ -5,7 +5,7 @@ import { Methods } from '@this/typey';
 
 import { DispatchInfo } from '#x/DispatchInfo';
 import { IncomingRequest } from '#x/IncomingRequest';
-import { OutgoingResponse } from '#x/OutgoingResponse';
+import { TypeOutgoingResponse } from '#x/TypeOutgoingResponse';
 
 
 /**
@@ -17,8 +17,8 @@ export class IntfRequestHandler {
   /**
    * Asks this instance to handle the given request; that is, parse it, act on
    * it, and either provide a response to send (or which could possibly be
-   * modified by an intermediary) or return `null` to indicate that the request
-   * was not handled.
+   * modified or replaced by an intermediary) or return `null` to indicate that
+   * the request was not handled.
    *
    * @abstract
    * @param {IncomingRequest} request Request object.
@@ -26,8 +26,7 @@ export class IntfRequestHandler {
    *   dispatch determination was made before calling this instance. (On any
    *   given instance -- depending on context -- it should be the case that it
    *   either _always_ or _never_ gets passed `null` for this parameter.)
-   * @returns {?OutgoingResponse} Response to send, or `null` if the request was
-   *   not in fact handled by this instance.
+   * @returns {?TypeOutgoingResponse} Response to send.
    * @throws {Error} Thrown in case of fatal error.
    */
   async handleRequest(request, dispatch) {